test: fix failing before all - should fail in CI
- combination of promise and cypress commands caused tests to be skipped
- moved helpers to e2e folder instead of cypress commands
