+++ cypress.json
@@ -2,6 +2,7 @@
     "viewportHeight": 600,
     "viewportWidth": 800,
     "video": false,
-    "integrationFolder": "test/e2e"
+    "integrationFolder": "test/e2e",
+    "testFiles": "**/*.e2e.js"
   }
