+++ cypress/plugins/index.js
@@ -26,4 +26,4 @@ module.exports = (on, config) => {
   // IMPORTANT to return the config object
   // with the any changed environment variables
   return config;
-}
+};
